+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 #### Next Release
 - Bug fixes:
   - Use internal file dir for sample app
+  - Use CopyOnWriteArrayList to store list of loggers
 - Updated dependencies:
   - Gradle to 6.6.1
   - Android Gradle plugin to 4.0.1